Update shortcodes.php
First off -- please forgive me for the terrible thing I have done here by mixing refactoring with code formatting :|

But! It's all worth it because the code is following WP standards and I've fixed a number of things so it's got some more flexibilty.

This includes the ability to set logged out text for the `[hidden_content]` shortcode and checking whether shortcodes exist before adding the tab/content for it in the `[account_tabs]` shortcode.

I initially wanted to get the tabs and the content separated so I could put the tabs into a sidebar widget but that wouldn't work without rewriting all of the JS and I decided against that.